fix: dependency cycles in public simulator - errors/revertReason
